+++ internal/app/handler/auth/function.go
