+++ internal/app/repository/auth/function.go
