+++ internal/app/usecase/plan/init.go
@@ -1,8 +1,8 @@
 package plan
 
 import (
+	"github.com/Marvit-Solutions/csw-golang/internal/app/repository/plan"
 	dto "github.com/Marvit-Solutions/csw-golang/internal/domain/response"
-	"github.com/Marvit-Solutions/csw-golang/internal/repository/plan"
 
 	"gorm.io/gorm"
 )
